% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst.R
\docType{data}
\name{dst}
\alias{dst}
\title{Pairwise comparisons of polygenic risk score development methods}
\format{
A data frame with 11374 rows and 35 columns:
\describe{
  \item{iy}{index for y (used internally while generating the table)}
  \item{beta_diff}{ \eqn{beta_x - beta_y} }
  \item{ix}{index for x (used internally while generating the table)}
  \item{se_x}{SE in the \link{metrics} table for method x}
  \item{beta_x}{BETA in the \link{metrics} table for method x}

  \item{se_y}{SE in the \link{metrics} table for method y}
  \item{beta_y}{BETA in the \link{metrics} table for method y}

  \item{cor_x_y}{correlation of scores x and y}
  \item{se_diff}{standard deviation of beta_diff, calculated as:
      \deqn{ \sqrt{ \text{se}_x^2 + \text{se}_y^2 - 2 \cdot \text{cor}_{xy} \cdot \text{se}_x \cdot \text{se}_y } }
  }
  \item{z_diff}{test statistic to calculate p_z_greater, p_z_smaller, and p_z_equal}
  \item{p_z_greater}{p-value for the z-test:
      \deqn{H_0: beta_y > beta_x} }
  \item{p_z_smaller}{p-value for the z-test:
      \deqn{ H_0: beta_y < beta_y} }
  \item{p_z_equal}{p-value for the z-test:
      \deqn{ H_0: beta_x - beta_y = 0} }
  \item{ci_high_diff}{ \deqn{ \text{beta}_\text{diff} + 1.96 \cdot \text{se}_\text{diff} } }
  \item{ci_low_diff}{ \deqn{ \text{beta}_\text{diff} - 1.96 \cdot \text{se}_\text{diff} } }

  The columns below have the same definitions as the \link{metrics} dataset:

  \item{n_x}{total sample size (number of cases + number of controls), method x}
  \item{n_y}{total sample size (number of cases + number of controls), method y}
  \item{h2l_x}{variance explained on the liability scale (this should be used instead of R2 for binary traits, based on R2_OBS), method x}
  \item{h2l_y}{variance explained on the liability scale (this should be used instead of R2 for binary traits, based on R2_OBS), method y}
  \item{r2_x}{variance explained on the observed scale (only really meaningful for continuous traits), method x}
  \item{r2_y}{variance explained on the observed scale (only really meaningful for continuous traits), method y}
  \item{method_x}{PRS method, method x}
  \item{method_y}{PRS method, method y}
  \item{method_type_x}{\strong{tuning type} (auto/CV), method x}
  \item{method_type_y}{\strong{tuning type} (auto/CV), method y}
  \item{tag_x}{similar to predictor, but uglier, method x}
  \item{tag_y}{similar to predictor, but uglier, method y}
  \item{phenotype}{the evaluated endpoint}
  \item{study}{the GWAS study identifier (GWAS catalog)}
  \item{ancestry}{the target ancestry}
  \item{bbid}{the target biobank}
  \item{ncas_x}{number of cases for binary endpoints, method x}
  \item{ncas_y}{number of cases for binary endpoints, method y}
  \item{ncon_x}{number of controls for binary endpoints, method x}
  \item{ncon_y}{number of controls for binary endpoints, method y}
  \item{auc_x}{AUROC for binary endpoints, method x}
  \item{auc_y}{AUROC for binary endpoints, method y}
}
}
\source{
\url{https://doi.org/10.1101/2023.11.20.23298215}
}
\usage{
dst
}
\description{
Pairwise comparisons of polygenic risk score development methods
}
\keyword{datasets}
