% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\docType{data}
\name{metrics}
\alias{metrics}
\title{Polygenic risk score performance predictive performance}
\format{
A data frame with 950 and 49 columns:
\describe{
  \item{predictor}{Country name}
  \item{N}{Explanation..}
  \item{h2l}{Explanation..}
  \item{BETA_CI_LOW, BETA_CI_HIGH}{2 & 3 letter ISO country codes}
  ...
}
}
\source{
<https://link/to/paper>
}
\usage{
metrics
}
\description{
A subset of data from... (science stuff here)
}
\keyword{datasets}
