% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_metrics.R
\name{plot_effect_sizes_full}
\alias{plot_effect_sizes_full}
\title{Title}
\usage{
plot_effect_sizes_full(
  plotdata,
  metric,
  plot_width,
  plot_height,
  legend_position = "bottom",
  vline = NA,
  plot_legend = F,
  dodge_width = 0.8,
  point_size = 0.3,
  nrow = 2,
  indicate_best = F,
  show_cv_equal_auto = T,
  scales = "fixed",
  shape = c(1, 16),
  x_variable = "method",
  y_lab = NA,
  strip_text_x_size = 12,
  axis_text_y_size = 12,
  axis_title_y_size = 20,
  adjust_methods_levels_fun = adjust_methods_levels
)
}
\arguments{
\item{adjust_methods_levels_fun}{}
}
\description{
Title
}
