% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prsCompaR-package.R
\docType{package}
\name{prsCompaR-package}
\alias{prsCompaR}
\alias{prsCompaR-package}
\title{prsCompaR: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
