% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotypes.R
\name{rename_phenotypes}
\alias{rename_phenotypes}
\title{Give phenotypes nice names}
\usage{
rename_phenotypes(x)
}
\arguments{
\item{x}{A data table containing a column called 'phenotype' or 'pheno'}
}
\value{
A data table with nicely named phenotypes
}
\description{
Give phenotypes nice names
}
